+++ spinetoolbox/executable_item_base.py
@@ -142,7 +142,7 @@ class ExecutableItemBase:
         return list()
 
     @classmethod
-    def from_dict(cls, item_dict, name, project_dir, app_settings, logger):
+    def from_dict(cls, item_dict, name, project_dir, app_settings, specifications, logger):
         """
         Deserializes an executable item from item dictionary.
 
@@ -151,6 +151,7 @@ class ExecutableItemBase:
             name (str): item's name
             project_dir (str): absolute path to the project directory
             app_settings (QSettings): Toolbox settings
+            specifications (dict): mapping from item specification name to :class:`ProjectItemSpecification`
             logger (LoggingInterface): a logger
         Returns:
             ExecutableItemBase: deserialized executable item