+++ spinetoolbox/project_items/importer/executable_item.py
@@ -16,6 +16,7 @@ Contains Importer's executable item as well as support utilities.
 :date:   1.4.2020
 """
 import os
+import pathlib
 from PySide2.QtCore import QObject, QEventLoop, Signal, Slot
 from spinetoolbox.config import PYTHON_EXECUTABLE
 from spinetoolbox.executable_item_base import ExecutableItemBase
@@ -23,6 +24,7 @@ from spinetoolbox.execution_managers import QProcessExecutionManager
 from spinetoolbox.spine_io.gdx_utils import find_gams_directory
 from . import importer_program
 from .item_info import ItemInfo
+from .utils import deserialize_mappings
 
 
 class ExecutableItem(ExecutableItemBase, QObject):
@@ -159,13 +161,24 @@ class ExecutableItem(ExecutableItemBase, QObject):
         self.importing_finished.emit()
 
     def _gams_system_directory(self):
-        """Returns GAMS system path from Toolbox settings or None if GAMS default is to be used."""
+        """Returns GAMS system path or None if GAMS default is to be used."""
         path = self._gams_path
         if not path:
             path = find_gams_directory()
         if path is not None and os.path.isfile(path):
             path = os.path.dirname(path)
         return path
+
+    @classmethod
+    def from_dict(cls, item_dict, name, project_dir, app_settings, specifications, logger):
+        """See base class."""
+        settings = deserialize_mappings(item_dict["mappings"], project_dir)
+        data_dir = pathlib.Path(project_dir, ".spinetoolbox", "items", item_dict["short name"])
+        logs_dir = os.path.join(data_dir, "logs")
+        python_path = app_settings.value("appSettings/pythonPath", defaultValue="")
+        gams_path = app_settings.value("appSettings/gamsPath", defaultValue=None)
+        cancel_on_error = item_dict["cancel_on_error"]
+        return cls(name, settings, logs_dir, python_path, gams_path, cancel_on_error, logger)
 
 
 def _files_from_resources(resources):